move network classes
