+++ app/src/main/java/com/pavel/a692group/data/datasource/remote/HttpGetRequestTask.java
@@ -1,4 +1,4 @@
-package com.pavel.a692group.request.http;
+package com.pavel.a692group.data.datasource.remote;
 
 import android.os.AsyncTask;
 import android.widget.TextView;
